requirements.txt
streamlit==1.28.1
rdkit==2023.9.5
pandas==2.1.1
pillow==10.0.1
numpy==1.24.3
